address link refactoring
