+++ src/components/ui/links/DisplayTransaction.tsx
@@ -2,7 +2,7 @@ import { HStack, Text, Tooltip } from '@chakra-ui/react';
 import { ArrowAngleUp } from '@decent-org/fractal-ui';
 import { useTranslation } from 'react-i18next';
 import { createAccountSubstring } from '../../../hooks/utils/useDisplayName';
-import EtherscanTransactionLink from './EtherscanTransactionLink';
+import EtherscanTransactionLink from './EtherscanLinkTransaction';
 
 export default function DisplayTransaction({ txHash }: { txHash: string }) {
   const displayName = createAccountSubstring(txHash);
