+++ src/components/ui/links/EtherscanLinkAddress.tsx
@@ -1,7 +1,11 @@
 import { Link } from '@chakra-ui/next-js';
-import { Copy } from '@decent-org/fractal-ui';
-import { useCopyText } from '../../../hooks/utils/useCopyText';
+import { LinkProps } from '@chakra-ui/react';
 import { useNetworkConfg } from '../../../providers/NetworkConfig/NetworkConfigProvider';
+
+interface Props extends LinkProps {
+  address?: string | null;
+  children: React.ReactNode;
+}
 
 /**
  * An address link to Etherscan.
@@ -9,44 +13,20 @@ import { useNetworkConfg } from '../../../providers/NetworkConfig/NetworkConfigP
  * For most use cases, you probably want to use DisplayAddress,
  * to add proper styling.
  */
-function EtherscanLinkAddress({
-  path = 'address',
-  address,
-  showCopyButton,
-  children,
-}: {
-  path?: string;
-  showCopyButton?: boolean;
-  address?: string | null;
-  children: React.ReactNode;
-}) {
+export default function EtherscanLinkAddress({ address, children, ...rest }: Props) {
   const { etherscanBaseURL } = useNetworkConfg();
-  const copyToClipboard = useCopyText();
 
   if (!address) {
     return null;
   }
 
-  const href = `${etherscanBaseURL}/${path}/${address}`;
-
   return (
-    <>
-      <Link
-        href={href}
-        target="_blank"
-      >
-        {children}
-      </Link>
-      {showCopyButton && (
-        <Copy
-          onClick={() => copyToClipboard(address)}
-          boxSize="1.5rem"
-          cursor="pointer"
-          ml={1}
-        />
-      )}
-    </>
+    <Link
+      href={`${etherscanBaseURL}/address/${address}`}
+      target="_blank"
+      {...rest}
+    >
+      {children}
+    </Link>
   );
 }
-
-export default EtherscanLinkAddress;
